requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
jinja2==3.1.2
pyyaml==6.0.1
